+++ examples/editor/src/PluginsConfig.js
@@ -7,6 +7,7 @@ import { LINK_TYPE } from 'wix-rich-content-plugin-link';
 import { VIDEO_TYPE } from 'wix-rich-content-plugin-video';
 import { GIPHY_TYPE } from 'wix-rich-content-plugin-giphy';
 import { MAP_TYPE } from 'wix-rich-content-plugin-map';
+import { FILE_UPLOAD_TYPE } from 'wix-rich-content-plugin-file-upload';
 import { EXTERNAL_MENTIONS_TYPE } from 'wix-rich-content-plugin-mentions';
 import { HEADERS_MARKDOWN_TYPE } from 'wix-rich-content-plugin-headers-markdown';
 import React from 'react';
@@ -194,6 +195,33 @@ export default {
       isViewControlShown: true,
       isDraggingAllowed: true,
     },
+  },
+  [FILE_UPLOAD_TYPE]: {
+    accept: '*',
+    onFileSelected: (file, updateEntity) => {
+      const name = file.name;
+      const filenameParts = name.split('.');
+      const type = filenameParts[filenameParts.length - 1];
+
+      const data = {
+        name,
+        type,
+        url: 'http://file-examples.com/wp-content/uploads/2017/10/file-sample_150kB.pdf',
+      };
+      setTimeout(() => updateEntity({ data }), 1000);
+    },
+    // handleFileSelection: updateEntity => {
+    //   const filenames = ['image.jpg', 'document.pdf', 'music.mp3'];
+    //   const name = filenames[Math.floor(Math.random() * filenames.length)];
+    //   const filenameParts = name.split('.');
+    //   const type = filenameParts[filenameParts.length - 1];
+    //   const data = {
+    //     name,
+    //     type,
+    //     url: 'http://file-examples.com/wp-content/uploads/2017/10/file-sample_150kB.pdf',
+    //   };
+    //   setTimeout(() => updateEntity({ data }), 500);
+    // },
   },
   uiSettings,
   getToolbarSettings: ({ pluginButtons, textButtons }) => [