+++ packages/plugin-image/src/image-viewer.jsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import includes from 'lodash/includes';
 import get from 'lodash/get';
 import isFunction from 'lodash/isFunction';
-import { mergeStyles, ImageLoader, validate } from 'wix-rich-content-common';
+import { mergeStyles, Loader, validate } from 'wix-rich-content-common';
 import isEqual from 'lodash/isEqual';
 import getImageSrc from './get-image-source';
 import { WIX_MEDIA_DEFAULT } from './get-wix-media-url';
@@ -131,7 +131,7 @@ class ImageViewer extends React.Component {
     }
     return (
       <div className={this.styles.imageOverlay}>
-        <ImageLoader type={'medium'} />
+        <Loader type={'medium'} />
       </div>
     );
   }